+++ app/Validators/StockTypeSchema.ts
@@ -1,0 +1,5 @@
+import { schema } from '@ioc:Adonis/Core/Validator'
+
+export const StockTypeSchema = schema.create({
+  stock_type: schema.string.optional(),
+})